% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_to_crosstalk.R
\name{calib_to_crosstalk}
\alias{calib_to_crosstalk}
\title{calib_to_crosstalk}
\usage{
calib_to_crosstalk(x)
}
\description{
This function extracts the direct crosstalk matrix from calibration data.
}
