% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossTalk.R
\name{CrossTalk}
\alias{CrossTalk}
\title{CrossTalk
Performs a CrossTalk analysis.}
\usage{
CrossTalk(
  cartridge_list,
  output_path,
  ch0 = 10,
  ch1 = 20,
  ch2 = 20,
  ch3 = 10,
  ch4 = 10,
  ch5 = 20,
  All = FALSE,
  measure = 1
)
}
\arguments{
\item{cartridge_list:}{dataframe containing the id and the RCA calls in an column called call. In long format.}

\item{output_path:}{file.path where the csv are located. (id.csv)}

\item{ch0:}{Integration time for Sensor 0.}

\item{ch1:}{Integration time for Sensor 1.}

\item{ch2:}{Integration time for Sensor 2.}

\item{ch3:}{Integration time for Sensor 3.}

\item{ch4:}{Integration time for Sensor 4.}

\item{ch5:}{Integration time for Sensor 5.}

\item{All:}{Boolean. Should all co-amplifications be computed?}

\item{measure:}{Select measure (desc. order)}
}
\description{
CrossTalk
Performs a CrossTalk analysis.
}
