% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSV_download.R
\name{CSV_download}
\alias{CSV_download}
\title{CSV_download
Downloads CSVs from in2data using the following inputs:}
\usage{
CSV_download(
  cartridges,
  in2data.url = "https://in2data.qiagen.com/in2data/backend/api/",
  use_cache = TRUE
)
}
\arguments{
\item{cartridges:}{List containing the id to use}

\item{path:}{file.path() of the output folder}

\item{in2data.url:}{url of the in2data API}

\item{use_cache:}{Boolean determining whether the cache will be used or not}
}
\description{
CSV_download
Downloads CSVs from in2data using the following inputs:
}
