% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCartridgeAM.R
\name{generateCartridgeAM}
\alias{generateCartridgeAM}
\title{generateCartridgeOM}
\usage{
generateCartridgeAM(
  am,
  in2data.url = "https://in2data.qiagen.com/in2data/backend/api/",
  fromTime = "1980-01-01"
)
}
\description{
This function returns the metadata for an AM under the UIA db format.
}
