% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCartridgeOM.R
\name{generateCartridgeOM}
\alias{generateCartridgeOM}
\title{generateCartridgeOM}
\usage{
generateCartridgeOM(
  om,
  in2data.url = "https://in2data.qiagen.com/in2data/backend/api/",
  fromTime = "1980-01-01"
)
}
\description{
This function returns the metadata for an OM under the UIA db format.
}
