% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCartridgeSID.R
\name{generateCartridgeSID}
\alias{generateCartridgeSID}
\title{generateCartridgeSID
This function returns the metadata for a set of cartridges based on the
SampleID. The cartridge list should be a dataframe where $testID is the name
of the variable to cycle.}
\usage{
generateCartridgeSID(
  list,
  in2data.url = "https://in2data.qiagen.com/in2data/backend/api/",
  fromTime = "1980-01-01"
)
}
\description{
generateCartridgeSID
This function returns the metadata for a set of cartridges based on the
SampleID. The cartridge list should be a dataframe where $testID is the name
of the variable to cycle.
}
