% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCartridgeTime.R
\name{generateCartridgeTime}
\alias{generateCartridgeTime}
\title{generateCartridgeTime
This function returns the metadata for a set of cartridges based on the
date.}
\usage{
generateCartridgeTime(
  fromTime = 0,
  in2data.url = "https://in2data.qiagen.com/in2data/backend/api/"
)
}
\description{
generateCartridgeTime
This function returns the metadata for a set of cartridges based on the
date.
}
