% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOMId.R
\name{generateOMId}
\alias{generateOMId}
\title{generateOMId}
\usage{
generateOMId(
  om,
  in2data.url = "https://in2data.qiagen.com/in2data/backend/api/"
)
}
\description{
This function returns the in2data id for a set list of OMs.
}
