% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateToken.R
\name{generateToken}
\alias{generateToken}
\title{generateToken}
\usage{
generateToken(
  endpoint = "https://in2data.qiagen.com/in2data/backend/api/",
  user = "api.merida",
  pass = "Api.2019.a6289577-20b9-4150-a488-fbe95e8346ad"
)
}
\description{
This function generates the token needed for the autentication to in2data.
}
