% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calibration.R
\name{get_calibration}
\alias{get_calibration}
\title{get_calibration
Returns a dataframe containing the calibration values for a set of IDs:}
\usage{
get_calibration(
  cartridges,
  in2data.url = "https://in2data.qiagen.com/in2data/backend/api/"
)
}
\arguments{
\item{cartridges:}{Dataframe containing the id to use}

\item{in2data.url:}{url of the in2data API}
}
\description{
get_calibration
Returns a dataframe containing the calibration values for a set of IDs:
}
