% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_techlog.R
\name{get_techlog}
\alias{get_techlog}
\title{get_techlog
Returns a dataframe containing the techlog values for a set of IDs:}
\usage{
get_techlog(
  cartridges,
  in2data.url = "https://in2data.qiagen.com/in2data/backend/api/"
)
}
\arguments{
\item{cartridges:}{Dataframe containing the id to use}

\item{in2data.url:}{url of the in2data API}
}
\description{
get_techlog
Returns a dataframe containing the techlog values for a set of IDs:
}
