% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCA_batch_regression.R
\name{RCA_batch_regression}
\alias{RCA_batch_regression}
\title{RCA_batch_regression
This performs a batch regression using the same inputs as RCA_LOAD_TEST
 but with a df}
\usage{
RCA_batch_regression(df, output = "All", cache = FALSE, filename = NULL)
}
\arguments{
\item{df:}{Dataframe containing: id of the cartridge, output folder, RCA_config, RCA_file, and RCA_conf_mgmt (or RCA, harness, calibration, config, barcode, and ignoreISflag). If a column is named 'expected' the concordance will be computed.}

\item{output:}{If the output is 'UIA' the output will be an unnested dataframe ready for FUTURO. 'All' to have an output as a nested Dataframe containing amplifications, targets, fails, and messages.}

\item{cache:}{Boolean. Should the output be stored?}

\item{filename:}{Path of the output.}
}
\description{
RCA_batch_regression
This performs a batch regression using the same inputs as RCA_LOAD_TEST
 but with a df
}
\examples{
c <- `Regression_ME_FPs` \%>\%
select(testID, 'group' = 'Group', 'expected' = 'Expected.RCA')

con <- RMySQL::dbConnect(RMySQL::MySQL(),
                        dbname = 'FuturoMilenial',
                        user = '*****',
                        password = '*****',
                        host = 'DB02STAT',
                        port = 3306)

cartridges_tbl <- tbl(con, 'Cartridges')

data <- cartridges_tbl \%>\% 
 select(id, uid) \%>\% 
 filter(uid \%in\% !!c$testID) \%>\% 
 collect \%>\% 
 mutate(out = "C:\\\\Users\\\\jmmerida\\\\Downloads\\\\test\\\\",
        csv_folder = "\\\\\\\\10.156.1.39\\\\cartridge_files",
        RCA_config = 'rca_men_bundle_2.0.1_NobN',
        RCA_file = file.path(file.path('\\\\\\\\10.104.48.55', 'RCAExplorer'), 'RCA'),
        RCA_conf_mgmt = file.path('RCAConfigurationManagement.csv')) \%>\% 
 left_join(c,
           by = c('uid' = 'testID'))

RMySQL::dbDisconnect(con)

a <- RCA_batch_regression(data,
                         cache = TRUE,
                         out = 'Theia',
                         filename = "C:\\\\Users\\\\jmmerida\\\\Downloads\\\\test\\\\test\\\\test.zip")

}
