% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCALoadTest.R
\name{RCA_load_test}
\alias{RCA_load_test}
\title{RCA.load.test
This performs a regression using the following inputs:}
\usage{
RCA_load_test(
  testID,
  out,
  csv_folder,
  RCA_config,
  RCA_file = NULL,
  RCA_conf_mgmt = NULL,
  RCA = NULL,
  harness = NULL,
  calibration = NULL,
  config = NULL,
  barcode = NULL,
  ignoreISflag = " -i"
)
}
\arguments{
\item{RCA_file}{: path of the RCA java file}

\item{testID:}{SN of the cartridge to execute}

\item{out:}{path of the output folder}

\item{RCA_config:}{chr. Bundle to use}

\item{csv_folder:}{path of the \*.csv files}

\item{RCA_conf_mgmt:}{LUT of RCA bundles}

\item{RCA:}{RCA jar file}

\item{harness:}{harness jar file}

\item{calibration:}{calibration csv}

\item{config:}{config xml}

\item{barcode:}{carcode txt}
}
\description{
RCA.load.test
This performs a regression using the following inputs:
}
