\name{RCAutils-package}
\alias{RCAutils-package}
\alias{RCAutils}
\docType{package}
\title{
\packageTitle{RCAutils}
}
\description{
\packageDescription{RCAutils}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RCAutils}
\packageIndices{RCAutils}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{RCAutils}

Maintainer: \packageMaintainer{RCAutils}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
