% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TechLog_download.R
\name{TechLog_download}
\alias{TechLog_download}
\title{TechLog_download
Downloads TechLogs from in2data using the following inputs:}
\usage{
TechLog_download(
  cartridges,
  in2data.url = "https://in2data.qiagen.com/in2data/backend/api/",
  use_cache = TRUE
)
}
\arguments{
\item{cartridges:}{List containing the id to use, the path, and the name of the file}

\item{in2data.url:}{url of the in2data API}

\item{use_cache:}{Boolean determining whether the cache will be used or not}
}
\description{
TechLog_download
Downloads TechLogs from in2data using the following inputs:
}
