% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateToken.R
\name{generateToken}
\alias{generateToken}
\title{generateToken}
\usage{
generateToken(
  endpoint = "https://in2data.qiagen.com/in2data/backend/api/",
  user = "api.delpozo",
  pass = "Api.2019.a20d487a-e05a-4499-b98e-705bf0144124"
)
}
\description{
This function generates the token needed for the autentication to in2data.
}
